requirements.txt
Django==4.2.7
tensorflow>=2.15.0
opencv-python
numpy
pandas
scikit-learn
matplotlib
Pillow
django-crispy-forms
crispy-bootstrap5